+++ sia/belief/gaussian.cpp
@@ -3,9 +3,9 @@
 
 #include "sia/belief/gaussian.h"
 #include "sia/common/exception.h"
+#include "sia/common/logger.h"
 #include "sia/math/math.h"
 
-#include <glog/logging.h>
 #include <cmath>
 
 namespace sia {
@@ -79,8 +79,8 @@ bool Gaussian::devectorize(const Eigen::VectorXd& data) {
   std::size_t n = dimension();
   std::size_t d = data.size();
   if (d != n * (n + 1)) {
-    LOG(WARNING) << "Devectorization failed, expected vector size "
-                 << n * (n + 1) << ", received " << d;
+    SIA_WARN("Devectorization failed, expected vector size "
+             << n * (n + 1) << ", received " << d);
     return false;
   }
   setMean(data.head(n));