+++ sia/controllers/mppi.cpp
@@ -2,8 +2,7 @@
 /// Licensed under BSD-3 Clause, https://opensource.org/licenses/BSD-3-Clause
 
 #include "sia/controllers/mppi.h"
-
-#include <glog/logging.h>
+#include "sia/common/logger.h"
 
 namespace sia {